+++ GHEtool/gui/translation_class.py
@@ -46,7 +46,6 @@ class Translations:
         "label_next",
         "label_previous",
         "hint_depth",
-        "option_show_legend",
         "function_save_results",
         "function_save_figure",
         "X_Axis",
@@ -193,6 +192,27 @@ class Translations:
         "option_method_temp_gradient",
         "option_method_rb_calc",
         "category_earth",
+        "cat_no_result",
+        "text_no_result",
+        "numerical_results",
+        "result_text_depth",
+        "result_Rb_calculated",
+        "results_ground_temperature",
+        "results_heating_load",
+        "results_heating_load_percentage",
+        "results_heating_ext",
+        "results_heating_peak",
+        "results_cooling_load",
+        "results_cooling_load_percentage",
+        "results_cooling_ext",
+        "results_cooling_peak",
+        "max_temp",
+        "min_temp",
+        "figure_temperature_profile",
+        "legend_figure_temperature_profile",
+        "hourly_figure_temperature_profile",
+        "figure_load_duration",
+        "legend_figure_load_duration",
         "languages",
     )
 
@@ -260,7 +280,7 @@ class Translations:
         self.page_aim: List[str] = [
             "Aim,Aim of simulation",
             "Ziel,Ziel der Simulation",
-            "Aim,Aim of simulation",
+            "Doel,Doel van de simulatie",
             "Aim,Aim of simulation",
             "Aim,Aim of simulation",
             "Aim,Aim of simulation",
@@ -269,7 +289,7 @@ class Translations:
         self.page_borehole: List[str] = [
             "Borehole @and earth,Borehole and earth",
             "Bohrloch @und Erdreich,Bohrloch und Erdreich",
-            "Boorveld @ en grond,Boorveld en grond",
+            "Boorveld @en grond,Boorveld en grond",
             "Foro e @terra,Foro e terra",
             "Forage @et terre,Forage et terre",
             "Pozo @y terreno,Pozo y terreno",
@@ -305,7 +325,7 @@ class Translations:
         self.page_options: List[str] = [
             "Options,Options",
             "Optionen,Optionen",
-            "Options,Options",
+            "Opties,Opties",
             "Options,Options",
             "Options,Options",
             "Options,Options",
@@ -344,7 +364,7 @@ class Translations:
         self.option_spacing: List[str] = [
             "Borehole spacing [m]: ",
             "Bohrlochabstand [m]: ",
-            "Boorgatspatiring [m]: ",
+            "Boorgatspatiëring [m]: ",
             "Spaziatura del foro [m]: ",
             "Espacement des trous de forage [m]: ",
             "Espaciado entre pozos [m]: ",
@@ -416,7 +436,7 @@ class Translations:
         self.option_heat_capacity: List[str] = [
             "Ground volumetric heat capacity [kJ / m³ K]:",
             "Spezifische Wärmekapazität des Erdreiches [kJ / m³ K]:",
-            "Ground volumetric heat capacity [kJ / m³ K]:",
+            "Volumetrische warmtecapaciteit van de grond [kJ / m³ K]:",
             "Ground volumetric heat capacity [kJ / m³ K]:",
             "Ground volumetric heat capacity [kJ / m³ K]:",
             "Ground volumetric heat capacity [kJ / m³ K]:",
@@ -470,7 +490,7 @@ class Translations:
         self.option_temp_gradient: List[str] = [
             "Temperature gradient [K/100m]: ",
             "Temperaturgradient [K/100m]: ",
-            "Temperature gradient [K/100m]: ",
+            "Temperatuursgradiënt [K/100m]: ",
             "Temperature gradient [K/100m]: ",
             "Temperature gradient [K/100m]: ",
             "Temperature gradient [K/100m]: ",
@@ -504,15 +524,6 @@ class Translations:
             "Profondeur du trou de sonde: ",
             "Profundidad del pozo: ",
             "Profundidade do pozo: ",
-        ]
-        self.option_show_legend: List[str] = [
-            "Show legend?,yes,no",
-            "Legende zeigen?,Ja,Nein",
-            "Toon legende?,yes,no",
-            "Mostra la legenda?,yes,no",
-            "Afficher la légende?,yes,no",
-            "Mostrar leyenda?,yes,no",
-            "Mostrar lenda?,yes,no",
         ]
         self.function_save_results: List[str] = [
             "Save results",
@@ -609,7 +620,7 @@ class Translations:
         self.hint_load_heating: List[str] = [
             "Heating load",
             "Heizlast",
-            "Belasting verwarming",
+            "Verwarmingslast",
             "Carico di riscaldamento",
             "Charge de chauffage",
             "Carga de calefacción",
@@ -618,7 +629,7 @@ class Translations:
         self.hint_load_cooling: List[str] = [
             "Cooling load",
             "Kühllast",
-            "Belasting koeling",
+            "Koellast",
             "Carico di raffreddamento",
             "Charge de refroidissement",
             "Carga de refrigeración",
@@ -685,7 +696,7 @@ class Translations:
         self.option_decimal_csv: List[str] = [
             "Decimal sign in CSV-file:,Point '.',Comma '++'",
             "Dezimalzeichen in CSV-Datei:,Punkt '.',Komma '++'",
-            "Decimaal teken in de CSV-file:,Punt '.',Komma '++'",
+            "Decimaalteken in de CSV-file:,Punt '.',Komma '++'",
             "Decimal sign in CSV-file:,Point '.',Comma '++'",
             "Decimal sign in CSV-file:,Point '.',Comma '++'",
             "Decimal sign in CSV-file:,Point '.',Comma '++'",
@@ -757,7 +768,7 @@ class Translations:
         self.option_column: List[str] = [
             "Thermal demand in one or two columns?:,1 column,2 columns",
             "Thermischer Bedarf einer oder zwei Spalten?:,1 Spalte,2 Spalten",
-            "Thermal demand in one or two columns?:,1 kolom,2 kolommen",
+            "Thermische vraag in één of twee kolommen?:,1 kolom,2 kolommen",
             "Thermal demand in one or two columns?:,1 colonna,2 colonne",
             "Thermal demand in one or two columns?;,1 colonne,2 colonnes",
             "Thermal demand in one or two columns?:,1 columna,2 columnas",
@@ -813,7 +824,7 @@ class Translations:
         self.ValueError: List[str] = [
             "Value error: check selected columns",
             "Wertefehler: ausgewählte Spalten prüfen",
-            "Value error: controleer geselecteerde kolommen",
+            "Waarde-error: controleer geselecteerde kolommen",
             "Errore di valore: controlla le colonne selezionate",
             "Erreur de valeur : vérifiez les colonnes sélectionnées",
             "Error de valor: compruebe las columnas seleccionadas",
@@ -822,7 +833,7 @@ class Translations:
         self.ColumnError: List[str] = [
             "Wrong column: check selected columns",
             "Falsche Spalte: ausgewählte Spalten prüfen",
-            "Wrong column: controleer geselecteerde kolommen",
+            "Foute kolom: controleer geselecteerde kolommen",
             "Colonna errata: controlla le colonne selezionate",
             "Colonne incorrecte : vérifiez les colonnes sélectionnées",
             "Columna incorrecta: compruebe las columnas seleccionadas",
@@ -1021,7 +1032,7 @@ class Translations:
         self.Y_Axis_Load_Q: List[str] = [
             "Remaining thermal energy [kWh]",
             "Übriggebliebene thermische Last [kWh]",
-            "Overblijvende energie energy [kWh]",
+            "Overblijvende thermische energie [kWh]",
             "Remaining thermal energy [kWh]",
             "Remaining thermal energy [kWh]",
             "Remaining thermal energy [kWh]",
@@ -1174,8 +1185,8 @@ class Translations:
         ]
         self.option_toggle_buttons: List[str] = [
             "Use toggle buttons?:, no , yes ",
-            "Umschalterbutton?:, no , yes ",
-            "Toggle buttons?:, no , yes ",
+            "Umschalterbutton?:, Ja , Nein ",
+            "Toggle-gedrag?:, ja , nee ",
             "Toggle buttons?:, no , yes ",
             "Toggle buttons?:, no , yes ",
             "Toggle buttons?:, no , yes ",
@@ -1184,7 +1195,7 @@ class Translations:
         self.option_auto_saving: List[str] = [
             "Use automatic saving?, no , yes ",
             "Automatisches speichern nutzen?, Nein, Ja ",
-            "Automatisch opslaan, no , yes ",
+            "Automatisch opslaan, nee , ja ",
             "Automatic saving, no , yes ",
             "Automatic saving, no , yes ",
             "Automatic saving, no , yes ",
@@ -1239,7 +1250,7 @@ class Translations:
         self.option_fluid_capacity: List[str] = [
             "Thermal capacity [J/kg K]:",
             "Wärmekapazität [J/kg K]:",
-            "Thermalisch warmtecapaciteit [J/kg K]:",
+            "Thermisch warmtecapaciteit [J/kg K]:",
             "Thermal capacity [J/kg K]:",
             "Thermal capacity [J/kg K]:",
             "Thermal capacity [J/kg K]:",
@@ -1421,7 +1432,7 @@ class Translations:
         self.aim_temp_profile: List[str] = [
             "Determine temperature profile",
             "Temperaturprofil berechnen",
-            "Determine temperature profile",
+            "Bepaal temperatuursprofiel",
             "Determine temperature profile",
             "Determine temperature profile",
             "Determine temperature profile",
@@ -1430,7 +1441,7 @@ class Translations:
         self.aim_req_depth: List[str] = [
             "Determine required depth",
             "Notwendige Tiefe berechnen",
-            "Determine required depth",
+            "Bereken benodigde diepte",
             "Determine required depth",
             "Determine required depth",
             "Determine required depth",
@@ -1439,7 +1450,7 @@ class Translations:
         self.aim_size_length: List[str] = [
             "Size by length and width",
             "Bestimme Länge und Breite des Bohrfeldes",
-            "Size by length and width",
+            "Dimensioneer bij breedte en lengte",
             "Size by length and width",
             "Size by length and width",
             "Size by length and width",
@@ -1448,7 +1459,7 @@ class Translations:
         self.aim_optimize: List[str] = [
             "Optimize load profile",
             "Optimiere Lastprofil",
-            "Optimize load profile",
+            "Optimaliseer belanstingsprofiel",
             "Optimize load profile",
             "Optimize load profile",
             "Optimize load profile",
@@ -1457,7 +1468,7 @@ class Translations:
         self.category_calculation: List[str] = [
             "Calculation options",
             "Berechnungsoptionen",
-            "Calculation options",
+            "Berekeningsopties",
             "Calculation options",
             "Calculation options",
             "Calculation options",
@@ -1466,16 +1477,16 @@ class Translations:
         self.option_method_size_depth: List[str] = [
             "Method for size borehole depth:,  L2  ,  L3  ,  L4  ",
             "Methode zur Bohrlochtiefendimensionierung:,  L2  ,  L3  ,  L4  ",
-            "Method for size borehole depth:,  L2  ,  L3  ,  L5  ",
-            "Method for size borehole depth:,  L2  ,  L3  ,  L5  ",
-            "Method for size borehole depth:,  L2  ,  L3  ,  L5  ",
-            "Method for size borehole depth:,  L2  ,  L3  ,  L5  ",
-            "Method for size borehole depth:,  L2  ,  L3  ,  L5  ",
+            "Methode voor boorvelddimensionering:,  L2  ,  L3  ,  L4  ",
+            "Method for size borehole depth:,  L2  ,  L3  ,  L4  ",
+            "Method for size borehole depth:,  L2  ,  L3  ,  L4  ",
+            "Method for size borehole depth:,  L2  ,  L3  ,  L4  ",
+            "Method for size borehole depth:,  L2  ,  L3  ,  L4  ",
         ]
         self.option_method_size_length: List[str] = [
             "Method for size width and length:,  L2  ,  L3  ",
             "Methode für Längen- und Breitendimensionierung:,  L2  ,  L3  ",
-            "Method for size width and length:,  L2  ,  L3  ",
+            "Methode voor boorvelddimensionering:,  L2  ,  L3  ",
             "Method for size width and length:,  L2  ,  L3  ",
             "Method for size width and length:,  L2  ,  L3  ",
             "Method for size width and length:,  L2  ,  L3  ",
@@ -1484,7 +1495,7 @@ class Translations:
         self.option_method_temp_gradient: List[str] = [
             "Should a temperature gradient over depth be considered?:, no , yes ",
             "Soll ein Temperaturgradient berücksichtigt werden?:, Nein , Ja ",
-            "Should a temperature gradient over depth be considered?:, no , yes ",
+            "Moet een temperatuursgradiënt in rekening worden gebracht?:, nee , ja ",
             "Should a temperature gradient over depth be considered?:, no , yes ",
             "Should a temperature gradient over depth be considered?:, no , yes ",
             "Should a temperature gradient over depth be considered?:, no , yes ",
@@ -1493,7 +1504,7 @@ class Translations:
         self.option_method_rb_calc: List[str] = [
             "Borehole resistance calculation method:, constant , dynamic ",
             "Mehtode zur Bohrlochwiderstangsberechnung:, Konstant , Dynamisch ",
-            "Borehole resistance calculation method:, constant , dynamic ",
+            "Berekeningsmethode boorgatweerstand:, constant , dynamisch ",
             "Borehole resistance calculation method:, constant , dynamic ",
             "Borehole resistance calculation method:, constant , dynamic ",
             "Borehole resistance calculation method:, constant , dynamic ",
@@ -1502,9 +1513,182 @@ class Translations:
         self.category_earth: List[str] = [
             "Earth properties",
             "Erdeigenschaften",
+            "Grondeigenschappen",
             "Earth properties",
             "Earth properties",
             "Earth properties",
             "Earth properties",
-            "Earth properties",
+        ]
+        self.cat_no_result: List[str] = ["No results", "Keine Ergebnisse", "Geen resultaten", "No results", "No results", "No results", "No results"]
+        self.text_no_result: List[str] = [
+            "No results are yet calculated",
+            "Es wurden noch keine Ergebnisse berechnet",
+            "Er zijn nog geen resultaten berekend",
+            "No results are yet calculated",
+            "No results are yet calculated",
+            "No results are yet calculated",
+            "No results are yet calculated",
+        ]
+        self.numerical_results: List[str] = [
+            "Numerical results",
+            "Numerische Ergebnisse",
+            "Numerische resultaten",
+            "Numerical results",
+            "Numerical results",
+            "Numerical results",
+            "Numerical results",
+        ]
+        self.result_text_depth: List[str] = ["Depth: , m", "Tiefe: , m", "Diepte: , m", "Depth: , m", "Depth: , m", "Depth: , m", "Depth: , m"]
+        self.result_Rb_calculated: List[str] = [
+            "Equivalent borehole thermal resistance: , mK/W",
+            "Äquivalenter thermischer Bohrlochwiderstand: , mK/W",
+            "Equivalente boorgatweerstand: , mK/W",
+            "Equivalent borehole thermal resistance: , mK/W",
+            "Equivalent borehole thermal resistance: , mK/W",
+            "Equivalent borehole thermal resistance: , mK/W",
+            "Equivalent borehole thermal resistance: , mK/W",
+        ]
+        self.results_ground_temperature: List[str] = [
+            "Average ground temperature: , °C",
+            "Durchschnittliche Erdreichtemperatur , °C",
+            "Gemiddelde grondtemperatuur: , °C",
+            "Average ground temperature: , °C",
+            "Average ground temperature: , °C",
+            "Average ground temperature: , °C",
+            "Average ground temperature: , °C",
+        ]
+        self.results_heating_load: List[str] = [
+            "Heating load on the borefield: , kWh",
+            "Die Heizlast für das Bohrfeld beträgt: , kWh",
+            "Verwarmingsvraag op het boorveld: , kWh",
+            "Heating load on the borefield: , kWh",
+            "Heating load on the borefield: , kWh",
+            "Heating load on the borefield: , kWh",
+            "Heating load on the borefield: , kWh",
+        ]
+        self.results_heating_load_percentage: List[str] = [
+            "This is , % of the heating load",
+            "Das sind , % der Heizlast",
+            "Dit is , % van de verwarmingsvraag",
+            "This is , % of the heating load",
+            "This is , % of the heating load",
+            "This is , % of the heating load",
+            "This is , % of the heating load",
+        ]
+        self.results_heating_ext: List[str] = [
+            "Heating load external: , kWh",
+            "Die externe Heizlast beträgt: , kWh",
+            "Verwarmingsvraag extern: , kWh",
+            "Heating load external: , kWh",
+            "Heating load external: , kWh",
+            "Heating load external: , kWh",
+            "Heating load external: , kWh",
+        ]
+        self.results_heating_peak: List[str] = [
+            "with a peak of: , kW",
+            "mit einer Spitzenlast von , kW",
+            "met een piek van: , kW",
+            "with a peak of: , kW",
+            "with a peak of: , kW",
+            "with a peak of: , kW",
+            "with a peak of: , kW",
+        ]
+        self.results_cooling_load: List[str] = [
+            "Cooling load on the borefield: , kWh",
+            "Die Kühllast für das Bohrfeld beträgt: , kWh",
+            "Koelvraag op het boorveld: , kWh",
+            "Cooling load on the borefield: , kWh",
+            "Cooling load on the borefield: , kWh",
+            "Cooling load on the borefield: , kWh",
+            "Cooling load on the borefield: , kWh",
+        ]
+        self.results_cooling_load_percentage: List[str] = [
+            "This is , % of the cooling load",
+            "Das sind , % der Kühllast",
+            "Dit is , % van de koelvraag",
+            "This is , % of the cooling load",
+            "This is , % of the cooling load",
+            "This is , % of the cooling load",
+            "This is , % of the cooling load",
+        ]
+        self.results_cooling_ext: List[str] = [
+            "Cooling load external: , kWh",
+            "Die externe Kühllast beträgt: , kWh",
+            "Koelvraag extern: , kWh",
+            "Cooling load external: , kWh",
+            "Cooling load external: , kWh",
+            "Cooling load external: , kWh",
+            "Cooling load external: , kWh",
+        ]
+        self.results_cooling_peak: List[str] = [
+            "with a peak of: , kW",
+            "mit einer Spitzenlast von , kW",
+            "met een piek van: , kW",
+            "with a peak of: , kW",
+            "with a peak of: , kW",
+            "with a peak of: , kW",
+            "with a peak of: , kW",
+        ]
+        self.max_temp: List[str] = [
+            "The maximum average fluid temperature is , °C",
+            "Die maximale Fluidtemperatur ist , °C",
+            "De maximaal gemiddelde fluïdumtemperatuur is , °C",
+            "The maximum average fluid temperature is , °C",
+            "The maximum average fluid temperature is , °C",
+            "The maximum average fluid temperature is , °C",
+            "The maximum average fluid temperature is , °C",
+        ]
+        self.min_temp: List[str] = [
+            "The minimal average fluid temperature is , °C",
+            "Die minimale Fluidtemperatur ist , °C",
+            "De minimaal gemiddelde fluïdumtemperatuur is , °C",
+            "The minimal average fluid temperature is , °C",
+            "The minimal average fluid temperature is , °C",
+            "The minimal average fluid temperature is , °C",
+            "The minimal average fluid temperature is , °C",
+        ]
+        self.figure_temperature_profile: List[str] = [
+            "Temperature evolution, Save figure",
+            "Temperaturverlauf, Abbildung speichern",
+            "Temperatuur evolutie, Sla figuur op",
+            "Temperature evolution, Save figure",
+            "Temperature evolution, Save figure",
+            "Temperature evolution, Save figure",
+            "Temperature evolution, Save figure",
+        ]
+        self.legend_figure_temperature_profile: List[str] = [
+            "Show legend?, No , Yes ",
+            "Legende zeigen?, Nein , Ja ",
+            "Toon legende?, no , yes ",
+            "Mostra la legenda?, no , yes ",
+            "Afficher la légende?, no , yes ",
+            "Mostrar leyenda?, no , yes ",
+            "Mostrar lenda?, no , yes ",
+        ]
+        self.hourly_figure_temperature_profile: List[str] = [
+            "Hourly profile, no , yes ",
+            "Stündliches Profil, Nein , Ja ",
+            "Uurlijks profiel, nee , ja ",
+            "Hourly profile, no , yes ",
+            "Hourly profile, no , yes ",
+            "Hourly profile, no , yes ",
+            "Hourly profile, no , yes ",
+        ]
+        self.figure_load_duration: List[str] = [
+            "Load-duration curve, Save figure",
+            "Jahresdauerlinie, Abbildung speichern",
+            "Belastings-duurcurve, Sla figuur op",
+            "Load-duration curve, Save figure",
+            "Load-duration curve, Save figure",
+            "Load-duration curve, Save figure",
+            "Load-duration curve, Save figure",
+        ]
+        self.legend_figure_load_duration: List[str] = [
+            "Show legend?, no , yes ",
+            "Legende zeigen?, Nein , Ja ",
+            "Toon legende?, nee , ja ",
+            "Mostra la legenda?, no , yes ",
+            "Afficher la légende?, no , yes ",
+            "Mostrar leyenda?, no , yes ",
+            "Mostrar lenda?, no , yes ",
         ]