+++ GHEtool/test/test_examples.py
@@ -19,6 +19,12 @@ def test_effect_borehole_configuration(monkeypatch):
     import GHEtool.Examples.effect_of_borehole_configuration
 
 
+@pytest.mark.slow
+def test_sizing_with_Rb(monkeypatch):
+    monkeypatch.setattr(plt, 'show', lambda: None)
+    import GHEtool.Examples.sizing_with_Rb_calculation
+
+
 def test_optimise_load_profile(monkeypatch):
     monkeypatch.setattr(plt, 'show', lambda: None)